+++ Sources/Extensions/SwiftLazy.swift
@@ -9,7 +9,14 @@
 import SwiftLazy
 
 extension Lazy: DelayMaker {
-  convenience init(_ factory: @escaping () -> Any?) {
+  public convenience init(file: String = #file, line: UInt = #line) {
+    self.init { () -> Value in
+      let name = (file as NSString).lastPathComponent
+      fatalError("Please inject this property from DI in file: \(name) on line: \(line). Lazy type: \(Value.self) ")
+    }
+  }
+
+  convenience init(_ container: DIContainer, _ factory: @escaping () -> Any?) {
     self.init { () -> Value in
       return gmake(by: factory())
     }
@@ -21,7 +28,14 @@ extension Lazy: DelayMaker {
 }
 
 extension Provider: DelayMaker {
-  convenience init(_ factory: @escaping () -> Any?) {
+  public convenience init(file: String = #file, line: UInt = #line) {
+    self.init { () -> Value in
+      let name = (file as NSString).lastPathComponent
+      fatalError("Please inject this property from DI in file: \(name) on line: \(line). Provider type: \(Value.self) ")
+    }
+  }
+
+  convenience init(_ container: DIContainer, _ factory: @escaping () -> Any?) {
     self.init { () -> Value in
       return gmake(by: factory())
     }
@@ -32,3 +46,93 @@ extension Provider: DelayMaker {
   }
 }
 
+// Providers with args
+extension Provider1: DelayMaker {
+  public convenience init(file: String = #file, line: UInt = #line) {
+    self.init { _ -> Value in
+      let name = (file as NSString).lastPathComponent
+      fatalError("Please inject this property from DI in file: \(name) on line: \(line). Provider type: \(Value.self) ")
+    }
+  }
+
+  static var type: DIAType { return Value.self }
+
+  convenience init(_ container: DIContainer, _ factory: @escaping () -> Any?) {
+    self.init { (arg1) -> Value in
+      container.extensions(for: Value.self)?.setArgs(arg1)
+      return gmake(by: factory())
+    }
+  }
+}
+
+extension Provider2: DelayMaker {
+  public convenience init(file: String = #file, line: UInt = #line) {
+    self.init { _, _ -> Value in
+      let name = (file as NSString).lastPathComponent
+      fatalError("Please inject this property from DI in file: \(name) on line: \(line). Provider type: \(Value.self) ")
+    }
+  }
+
+  static var type: DIAType { return Value.self }
+
+  convenience init(_ container: DIContainer, _ factory: @escaping () -> Any?) {
+    self.init { (arg1, arg2) -> Value in
+      container.extensions(for: Value.self)?.setArgs(arg1, arg2)
+      return gmake(by: factory())
+    }
+  }
+}
+
+extension Provider3: DelayMaker {
+  public convenience init(file: String = #file, line: UInt = #line) {
+    self.init { _, _, _ -> Value in
+      let name = (file as NSString).lastPathComponent
+      fatalError("Please inject this property from DI in file: \(name) on line: \(line). Provider type: \(Value.self) ")
+    }
+  }
+
+  static var type: DIAType { return Value.self }
+
+  convenience init(_ container: DIContainer, _ factory: @escaping () -> Any?) {
+    self.init { (arg1, arg2, arg3) -> Value in
+      container.extensions(for: Value.self)?.setArgs(arg1, arg2, arg3)
+      return gmake(by: factory())
+    }
+  }
+}
+
+extension Provider4: DelayMaker {
+  public convenience init(file: String = #file, line: UInt = #line) {
+    self.init { _, _, _, _ -> Value in
+      let name = (file as NSString).lastPathComponent
+      fatalError("Please inject this property from DI in file: \(name) on line: \(line). Provider type: \(Value.self) ")
+    }
+  }
+
+  static var type: DIAType { return Value.self }
+
+  convenience init(_ container: DIContainer, _ factory: @escaping () -> Any?) {
+    self.init { (arg1, arg2, arg3, arg4) -> Value in
+      container.extensions(for: Value.self)?.setArgs(arg1, arg2, arg3, arg4)
+      return gmake(by: factory())
+    }
+  }
+}
+
+extension Provider5: DelayMaker {
+  public convenience init(file: String = #file, line: UInt = #line) {
+    self.init { _, _, _, _, _ -> Value in
+      let name = (file as NSString).lastPathComponent
+      fatalError("Please inject this property from DI in file: \(name) on line: \(line). Provider type: \(Value.self) ")
+    }
+  }
+
+  static var type: DIAType { return Value.self }
+
+  convenience init(_ container: DIContainer, _ factory: @escaping () -> Any?) {
+    self.init { (arg1, arg2, arg3, arg4, arg5) -> Value in
+      container.extensions(for: Value.self)?.setArgs(arg1, arg2, arg3, arg4, arg5)
+      return gmake(by: factory())
+    }
+  }
+}